+++ docs/conf.py
@@ -16,6 +16,9 @@
 # absolute, like shown here.
 #
 
+import os
+import sys
+sys.path.insert(0, os.path.abspath('..'))
 
 # -- General configuration ---------------------------------------------