+++ tests/test_orm.py
@@ -1,4 +1,5 @@
 import json
+from typing import List, Dict
 
 from grpcalchemy.orm import (
     BooleanField,
@@ -36,14 +37,22 @@ class ORMTestCase(TestGrpcalchemy):
             name = StringField()
             number = Int32Field()
             big_number = Int64Field()
-            sex = BooleanField()
+            active = BooleanField()
             raw_data = BytesField()
 
-        self.assertEqual("string name", str(Test.name))
-        self.assertEqual("int32 number", str(Test.number))
         self.assertEqual("int64 big_number", str(Test.big_number))
-        self.assertEqual("bool sex", str(Test.sex))
-        self.assertEqual("bytes raw_data", str(Test.raw_data))
+
+        class TestWithTyping(Message):
+            name: str
+            number: int
+            active: bool
+            raw_data: bytes
+
+        for object in [Test, TestWithTyping]:
+            self.assertEqual("string name", str(object.name))
+            self.assertEqual("int32 number", str(object.number))
+            self.assertEqual("bool active", str(object.active))
+            self.assertEqual("bytes raw_data", str(object.raw_data))
 
     def test_ReferenceField(self):
         """Test normal and list reference field"""
@@ -52,41 +61,59 @@ class ORMTestCase(TestGrpcalchemy):
             name = StringField()
 
         class TestRef(Message):
-            ref_field = ReferenceField(Test)
-            list_test_field = ListField(Test)
-            list_int32_field = ListField(Int32Field)
+            ref_field = ReferenceField(Test())
+            list_test_field = ListField(ReferenceField(Test()))
+            list_int_field = ListField(Int64Field())
 
-        self.assertEqual([], TestRef().list_test_field)
-        self.assertEqual([], TestRef().list_int32_field)
+        class TestWithTyping(Message):
+            name: str
 
-        test = TestRef(
-            ref_field=Test(name="Test"),
-            list_test_field=[Test(name="Test")],
-            list_int32_field=[1],
-        )
-        self.assertEqual("Test", test.ref_field.name)
-        test.ref_field.name = "Changed_name"
-        self.assertEqual("Changed_name", test.ref_field.name)
-        self.assertListEqual([1], list(test.list_int32_field))
-        self.assertEqual("Test", test.list_test_field[0].name)
+        class TestRefWithTyping(Message):
+            ref_field: TestWithTyping
+            list_test_field: List[TestWithTyping]
+            list_int_field: List[int]
 
-        self.assertEqual("Test ref_field", str(TestRef.ref_field))
-        self.assertEqual("repeated Test list_test_field", str(TestRef.list_test_field))
-        self.assertEqual(
-            "repeated int32 list_int32_field", str(TestRef.list_int32_field)
-        )
+        for object in [TestRef, TestRefWithTyping]:
+            self.assertEqual([], object().list_test_field)
+            self.assertEqual([], object().list_int_field)
+
+            test = object(
+                ref_field=Test(name="Test"),
+                list_test_field=[Test(name="Test")],
+                list_int_field=[1],
+            )
+            self.assertEqual("Test", test.ref_field.name)
+            test.ref_field.name = "Changed_name"
+            self.assertEqual("Changed_name", test.ref_field.name)
+            self.assertListEqual([1], list(test.list_int_field))
+            self.assertEqual("Test", test.list_test_field[0].name)
+
+            self.assertEqual("Test ref_field", str(TestRef.ref_field))
+            self.assertEqual(
+                "repeated Test list_test_field", str(TestRef.list_test_field)
+            )
+            self.assertEqual(
+                "repeated int64 list_int_field", str(TestRef.list_int_field)
+            )
 
     def test_MapField(self):
         class Test(Message):
             name = StringField()
 
         class TestMapRef(Message):
-            map_field = MapField(StringField, Test)
+            map_field = MapField(StringField(), ReferenceField(Test()))
 
-        test = TestMapRef(map_field={"test": Test(name="test")})
-        self.assertEqual("test", test.map_field["test"].name)
-        self.assertEqual("test", test.__message__.map_field["test"].name)
-        self.assertEqual("map<string, Test> map_field", str(TestMapRef.map_field))
+        class TestWithTyping(Message):
+            name: str
+
+        class TestMapRefWithTyping(Message):
+            map_field: Dict[str, TestWithTyping]
+
+        for object in [TestMapRef, TestMapRefWithTyping]:
+            test = object(map_field={"test": Test(name="test")})
+            self.assertEqual("test", test.map_field["test"].name)
+            self.assertEqual("test", test.__message__.map_field["test"].name)
+            self.assertEqual("map<string, Test> map_field", str(TestMapRef.map_field))
 
     def test_inheritance(self):
         class Post(Message):
@@ -101,13 +128,13 @@ class ORMTestCase(TestGrpcalchemy):
 
     def test_json_format(self):
         class Test(Message):
-            name = StringField()
+            name: str
 
         class TestJsonFormat(Message):
-            ref_field = ReferenceField(Test)
-            list_test_field = ListField(Test)
-            list_int32_field = ListField(Int32Field)
-            map_field = MapField(StringField, Test)
+            ref_field: Test
+            list_test_field: List[Test]
+            list_int32_field: List[int]
+            map_field: Dict[str, Test]
 
         generate_proto_file()
 
@@ -129,17 +156,17 @@ class ORMTestCase(TestGrpcalchemy):
 
     def test_field_default_value(self):
         class Test(Message):
-            name = StringField()
-            number = Int32Field()
+            name: str
+            number: int
             big_number = Int64Field()
-            sex = BooleanField()
-            raw_data = BytesField()
+            active: bool
+            raw_data: bytes
 
         test = Test()
         self.assertEqual("", test.name)
         self.assertEqual(0, test.number)
         self.assertEqual(0, test.big_number)
-        self.assertEqual(False, test.sex)
+        self.assertEqual(False, test.active)
         self.assertEqual(b"", test.raw_data)
 
         test.name = "changed"